changed directory structure again ffs
